+++ CHANGELOG.md
@@ -19,8 +19,12 @@
 
 # REVISIONS
 # v0.8.12
-- Updated `Do` to `Issue`
-- 
+- [x] Updated `Do` to `Issue`
+- [x] Update `Compile` component to spit out a `RobotProgram` instance
+- [ ] Update `DisplayProgram` to render it 
+- [ ] Update `WriteProgram` to dump the RobotProgram and save it to a folder
+- [ ] Update sample files to reflect these changes (add entwine too).
+
 
 # v0.8.8
 - Core update